feat(checker): extract checker command
