+++ apps/checker/ping.go
@@ -1,4 +1,4 @@
-package main
+package checker
 
 import (
 	"bytes"
@@ -25,27 +25,7 @@ type PingData struct {
 	Message       string `json:"message,omitempty"`
 }
 
-func sendToTinybird(pingData PingData) {
-	url := "https://api.tinybird.co/v0/events?name=golang_ping_response__v3"
-	fmt.Printf("📈  Sending data to Tinybird for %+v \n", pingData)
-	bearer := "Bearer " + os.Getenv("TINYBIRD_TOKEN")
-	payloadBuf := new(bytes.Buffer)
-	json.NewEncoder(payloadBuf).Encode(pingData)
-	req, err := http.NewRequest("POST", url, payloadBuf)
-	req.Header.Set("Authorization", bearer)
-	req.Header.Set("Content-Type", "application/json")
-
-	client := &http.Client{Timeout: time.Second * 10}
-	_, err = client.Do(req)
-	if err != nil {
-		fmt.Println(err)
-		panic(err)
-	}
-	// Should we add a retry mechanism here?
-
-}
-
-func sendToTinybirdNew(pingData PingData) {
+func SendToTinyBird(pingData PingData) {
 	url := "https://api.tinybird.co/v0/events?name=ping_response__v5"
 	fmt.Printf("📈  Sending data to Tinybird for %+v \n", pingData)
 	bearer := "Bearer " + os.Getenv("TINYBIRD_TOKEN")
@@ -65,7 +45,7 @@ func sendToTinybirdNew(pingData PingData) {
 
 }
 
-func ping(client *http.Client, inputData request.CheckerRequest) (PingData, error) {
+func Ping(client *http.Client, inputData request.CheckerRequest) (PingData, error) {
 
 	region := os.Getenv("FLY_REGION")
 	request, err := http.NewRequest(inputData.Method, inputData.URL, bytes.NewReader([]byte(inputData.Body)))
